# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Friend Buyer API - inventory management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Me {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  fullName?: string | null;
  roles: string[];
  /** @nullable */
  storeId?: string | null;
  /** @nullable */
  warehouseId?: string | null;
}

export interface DashboardSummary {
  companies: number;
  stores: number;
  warehouses: number;
  suppliers: number;
  skus: number;
  unitsInStock: number;
}

export interface ActivityItem {
  id: string;
  movementType: string;
  /** @nullable */
  skuCode?: string | null;
  quantity: number;
  createdAt: string;
}

export interface Company {
  id: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface CompanyInput {
  name: string;
}

export interface Store {
  id: string;
  name: string;
  companyId: string;
  createdAt: string;
  updatedAt: string;
}

export interface StoreInput {
  name: string;
  companyId: string;
}

export interface Warehouse {
  id: string;
  name: string;
  storeId: string;
  createdAt: string;
  updatedAt: string;
}

export interface WarehouseInput {
  name: string;
  storeId: string;
}

export interface Supplier {
  id: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface SupplierInput {
  name: string;
}

export interface Sku {
  id: string;
  code: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  description?: string | null;
  isSerialized: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface SkuInput {
  code: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  description?: string | null;
  isSerialized?: boolean;
}