# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Friend Buyer API - inventory management
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  movementType: string;
  /** @nullable */
  skuCode?: string | null;
  quantity: number;
  createdAt: Date;
}