# lib/api-zod/src/generated/types/sku.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Friend Buyer API - inventory management
 * OpenAPI spec version: 0.1.0
 */

export interface Sku {
  id: string;
  code: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  description?: string | null;
  isSerialized: boolean;
  createdAt: Date;
  updatedAt: Date;
}