# lib/api-zod/src/generated/types/warehouse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Friend Buyer API - inventory management
 * OpenAPI spec version: 0.1.0
 */

export interface Warehouse {
  id: string;
  name: string;
  storeId: string;
  createdAt: Date;
  updatedAt: Date;
}