# lib/api-zod/src/generated/types/warehouseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Friend Buyer API - inventory management
 * OpenAPI spec version: 0.1.0
 */

export interface WarehouseInput {
  name: string;
  storeId: string;
}